+++ src/store/gifts.ts
@@ -2,9 +2,11 @@ import {
     LOOTBOX_TYPE_GIFTS,
     STATUSES_LOST_GIFT,
 } from "@src/config/gift";
-import { defineStore } from "pinia";
+import { defineStore, storeToRefs } from "pinia";
 import { computed, ref } from "vue";
 
+import { currencyView } from "../helpers/currencyHelper";
+import { Currencies } from "../models/enums/currencies";
 import type { GiftAllItem, IGift, IGiftDeposit, IGiftFreeSpins, IGiftModifyConfig } from "../services/api/DTO/gifts";
 import { useUserInfo } from "./user/userInfo";
 import { useUserStatuses } from "./user/userStatuses";
@@ -19,6 +21,7 @@ export const useGiftsStore = defineStore("giftsStore", () => {
     const modifyGiftsConfig = ref<IGiftModifyConfig[]>([]);
     const additionalGifts = ref({});
     const depositGiftsAll = ref<IGiftDeposit[]>([]);
+    const activeDepositGift = ref<IGiftDeposit | null>(null);
     const registrationGiftsAll = ref<IGiftDeposit[]>([]);
     const fsGiftsAll = ref<IGiftFreeSpins[]>([] as IGiftFreeSpins[]);
 
@@ -42,7 +45,7 @@ export const useGiftsStore = defineStore("giftsStore", () => {
 
     const depositGifts = computed<IGiftDeposit[]>(() => {
         const localGifts = depositGiftsAll.value.filter((gift) => {
-            return gift.bonuses?.[0]?.type && !LOOTBOX_TYPE_GIFTS.includes(gift.bonuses[0].type);
+            return gift.bonuses?.[0]?.type && !LOOTBOX_TYPE_GIFTS?.includes(gift.bonuses[0].type);
         });
 
         if (additionalGift.value) {
@@ -109,6 +112,95 @@ export const useGiftsStore = defineStore("giftsStore", () => {
     const giftsCounter = computed<number>(() => {
         return giftsAll.value.length;
     });
+
+
+    const activeDepositGiftGroupID = computed(() => {
+        if (activeDepositGift.value) {
+            return getDepositGiftGroupID(activeDepositGift.value);
+        }
+
+        return null;
+    });
+
+    const activeDepositGiftMinDep = computed(() => {
+        const { getUserCurrency } = storeToRefs(useUserInfo());
+        const { getSubunitsToUnitsByCode } = useUserInfo();
+
+        if (!activeDepositGift.value?.id) {
+            return 0;
+        }
+
+        const isLootbox = activeDepositGift.value.bonuses[0]?.type === "random" && activeDepositGift.value.bonuses[0].boxes;
+        let bonuses = null;
+
+        if (isLootbox) {
+            const boxes = activeDepositGift.value.bonuses[0].boxes;
+            bonuses = boxes?.[boxes.length - 1].bonuses[0];
+        } else {
+            bonuses = activeDepositGift.value.bonuses[0];
+        }
+
+        const attr = bonuses?.conditions.find((item) => {
+            return item.field === "amount" && item.type === "min";
+        });
+
+        if (attr) {
+            const valueAttr = attr.value.find((item) => {
+                return item.currency === getUserCurrency.value;
+            });
+
+            if (valueAttr) {
+                return currencyView(
+                    valueAttr.amount_cents,
+                    valueAttr.currency,
+                    null,
+                    getSubunitsToUnitsByCode(valueAttr.currency as Currencies),
+                    8,
+                    false,
+                );
+            }
+        }
+
+        return 0;
+    });
+
+    const giftMatchInUserGroup = computed(() => {
+        return depositGifts.value.find((gift: IGiftDeposit) => {
+            // @ts-expect-error No overload matches this call.
+            if (userStatuses.getUserGroups.includes(getDepositGiftGroupID(gift))) {
+                return gift;
+            }
+        });
+    });
+
+    const giftMatchInUserGroupID = computed(() => {
+        if (giftMatchInUserGroup.value) {
+            return getDepositGiftGroupID(giftMatchInUserGroup.value);
+        }
+
+        return null;
+    });
+
+    function getDepositGiftGroupID(gift: IGiftDeposit) {
+        const conditionId = gift?.bonuses?.[0]?.conditions
+            ?.find((condition) => condition.field === "groups")
+            ?.value?.find((item) => String(item).includes("pick:"));
+
+        return conditionId || null;
+    };
+
+    function setActiveDepositGift(value: IGiftDeposit | null) {
+        activeDepositGift.value = value;
+    }
+
+    async function resetActiveDepositGift() {
+        if (activeDepositGiftGroupID.value) {
+            // @ts-expect-error No overload matches this call.
+            await userStatuses.changeUserToGroup(null, activeDepositGiftGroupID.value);
+            setActiveDepositGift(null);
+        }
+    }
+
 
     function setGiftsLoading(value: boolean): void {
         isLoadingGiftData.value = value;
@@ -182,5 +274,14 @@ export const useGiftsStore = defineStore("giftsStore", () => {
 
         isLoadingGiftData,
         setGiftsLoading,
+
+        activeDepositGift,
+        activeDepositGiftGroupID,
+        activeDepositGiftMinDep,
+        giftMatchInUserGroup,
+        giftMatchInUserGroupID,
+        getDepositGiftGroupID,
+        setActiveDepositGift,
+        resetActiveDepositGift,
     };
 });