+++ src/store/user/userInfo.ts
@@ -159,6 +159,14 @@ export const useUserInfo = defineStore("userInfo", () => {
         info.value.statuses.push(userGroup);
     }
 
+    function removeUserGroup(userGroup: IUserStatus) {
+        const indexCurrentGroup = info.value.statuses.findIndex((status) => status.id === userGroup.id);
+
+        if (indexCurrentGroup >= 0) {
+            info.value.statuses.splice(indexCurrentGroup, 1);
+        }
+    }
+
     function setUserStatuses(userGroups: IUserStatus[]) {
         info.value.statuses = userGroups;
     }
@@ -180,7 +188,7 @@ export const useUserInfo = defineStore("userInfo", () => {
     function addUserStatuses(newStatus: IUserStatus) {
         info.value.statuses = [
             ...info.value.statuses.filter(({ id }) => {
-                return Number(id) !== Number(newStatus.id);
+                return id !== newStatus.id;
             }),
             newStatus,
         ];
@@ -414,6 +422,7 @@ export const useUserInfo = defineStore("userInfo", () => {
         toggleUserIsLogged,
         setUserData,
         addUserGroup,
+        removeUserGroup,
         clearUserData,
         addUserStatuses,
         setUserStatuses,