+++ src/store/user/userStatuses.ts
@@ -1,5 +1,6 @@
 import {
     ALL_LEVELS,
+    ID_CASHBOX_ONBOARD_DONE,
     ID_GROUP_FOR_MULTI_ACC,
     TEST_GROUP_ID,
 } from "@config/user-statuses";
@@ -12,7 +13,7 @@ import { computed, ref } from "vue";
 import type { IUserStatus, UserGroup } from "../../models/user";
 import { IVipManager } from "../../models/vipManagers";
 import { loadManagersConfigReq } from "../../services/api/requests/configs";
-import { addPlayerToGroup } from "../../services/api/requests/player";
+import { changePlayerGroup, type IPlayerGroup } from "../../services/api/requests/player";
 import { useLevelsStore } from "../levels/levelsStore";
 import { useUserInfo } from "./userInfo";
 
@@ -78,10 +79,21 @@ export const useUserStatuses = defineStore("userStatuses", () => {
         return userManager.value;
     });
 
-    async function addUserToGroup(groupForAdding: string | number) {
-        if (!getUserGroups.value.includes(groupForAdding)) {
-            await addPlayerToGroup(groupForAdding);
-            userStore.addUserGroup({ id: groupForAdding, name: "" });
+    const isCashboxOnboardDone = computed(() => {
+        return getUserGroups.value.includes(ID_CASHBOX_ONBOARD_DONE);
+    });
+
+    async function changeUserToGroup(groupForAdding?: IPlayerGroup, groupForRemoving?: IPlayerGroup) {
+        await changePlayerGroup(groupForAdding, groupForRemoving);
+
+        if (groupForAdding) {
+            if (!getUserGroups.value.includes(groupForAdding)) {
+                userStore.addUserGroup({ id: groupForAdding, name: groupForAdding });
+            }
+
+            if (getUserGroups.value.includes(groupForAdding) && groupForRemoving) {
+                userStore.removeUserGroup({ id: groupForRemoving, name: groupForRemoving });
+            }
         }
     }
 
@@ -109,13 +121,14 @@ export const useUserStatuses = defineStore("userStatuses", () => {
         isMultiAccount,
         isVip,
         isDiamond,
+        isCashboxOnboardDone,
         getUserManager,
         userVipStatus,
         userVipGroup,
         isRegisteredViaSocialNetwork,
         getUserLevelId,
+        changeUserToGroup,
         setSocialNetworkAuthGroups,
-        addUserToGroup,
         loadUserManager,
         clearUserManager,
     };